refactor(gateway): extract resume/replay helpers from run.py (slice 3 of #54962)
Third slice of the gateway god-file unpacking: extract
build_resume_recovery_note, _build_replay_entry and the shared
_ASSISTANT_REPLAY_FIELDS whitelist into
gateway/resume_replay_helpers.py.

- Byte-identical extraction (AST-verified against origin/main): no
  behavior change
- gateway/run.py: -172 lines; helpers imported at the extraction point
  (module-attribute imports keep existing tests green)
- 43 tests pass across replay-entry, resume-pending, and message
  timestamp suites

Follow-up to #77433 (slice 1) and #77438 (slice 2). Progress on #54962.
