+++ gateway/run.py
@@ -935,177 +935,12 @@ def _is_fresh_gateway_interruption(
     return current - timestamp <= window
 
 
-def build_resume_recovery_note(
-    reason: Optional[str],
-    message: str = "",
-    *,
-    interactive: bool = True,
-) -> str:
-    """Build the resume-pending recovery system note for an interrupted turn.
-
-    ``reason`` is the session's ``resume_reason`` (``restart_timeout``,
-    ``shutdown_timeout``, or anything else → generic interruption phrasing).
-    ``message`` is the user's NEW message text; empty means this is the
-    startup auto-resume turn synthesized by
-    ``_schedule_resume_pending_sessions`` with no human message attached.
-
-    ``interactive`` selects the empty-message guidance: on interactive
-    platforms a human is present, so "report the restore and ask what next"
-    is right.  On non-interactive event platforms (webhook, API server —
-    adapters with ``interactive_resume = False``) nobody can answer; the
-    resumed turn must instead complete the interrupted work, or the task is
-    silently abandoned behind a "restored" acknowledgement that goes
-    nowhere (#57056).
-    """
-    reason_phrase = (
-        "a gateway restart"
-        if reason == "restart_timeout"
-        else "a gateway shutdown"
-        if reason == "shutdown_timeout"
-        else "a gateway interruption"
-    )
-    if message:
-        resume_guidance = (
-            "Address the user's NEW message below FIRST and focus "
-            "on what the user is asking now."
-        )
-        tail_guidance = (
-            "Do NOT re-execute old tool calls — skip any "
-            "unfinished work from the conversation history."
-        )
-    elif interactive:
-        resume_guidance = (
-            "Report to the user that the session was restored "
-            "successfully and ask what they would like to do next."
-        )
-        tail_guidance = (
-            "Do NOT re-execute old tool calls — skip any "
-            "unfinished work from the conversation history."
-        )
-    else:
-        resume_guidance = (
-            "No user is present on this non-interactive platform, "
-            "so do NOT emit a 'session restored' acknowledgement "
-            "or ask questions. Review the conversation history and "
-            "CONTINUE the interrupted task to completion."
-        )
-        tail_guidance = (
-            "Do NOT re-run tool calls whose results already "
-            "appear in the history — resume from the first step "
-            "that has no recorded result."
-        )
-    return (
-        f"[System note: The previous turn was interrupted by "
-        f"{reason_phrase}; the gateway is now back online. "
-        f"Any restart/shutdown command in the history has already "
-        f"run — do NOT re-execute or verify it. {resume_guidance} "
-        f"{tail_guidance}]"
-        + (f"\n\n{message}" if message else "")
-    )
-
-
-# Assistant-message fields that must survive transcript replay so multi-turn
-# reasoning context, prefix-cache hits, and provider-specific echo
-# requirements all behave the same on the gateway as they do in the CLI.
-#
-# ``reasoning`` and ``reasoning_details`` were the original three preserved
-# by PR #2974 (schema v6).  ``reasoning_content``, ``codex_reasoning_items``,
-# ``codex_message_items``, and ``finish_reason`` were added to the DB later
-# but the gateway's replay whitelist was never expanded to match — so any
-# pure-text assistant turn (no ``tool_calls``) silently dropped them on
-# replay, regressing the CLI-vs-gateway behavioural parity.
-#
-# Why each field matters on replay:
-#   * ``reasoning`` / ``reasoning_content``: provider-facing thinking text.
-#     ``_copy_reasoning_content_for_api`` promotes ``reasoning`` →
-#     ``reasoning_content`` at send time, but only when the strings happen to
-#     match.  Carrying the original ``reasoning_content`` verbatim avoids
-#     reconstruction loss for providers that return them as distinct fields
-#     (DeepSeek/Kimi/Moonshot thinking modes).
-#   * ``reasoning_details``: opaque structured array (signature,
-#     encrypted_content) used by OpenRouter/Anthropic to maintain reasoning
-#     continuity across turns.
-#   * ``codex_reasoning_items``: encrypted reasoning blobs for the OpenAI
-#     Codex Responses API.
-#   * ``codex_message_items``: exact assistant message items with ``phase``.
-#     OpenAI docs: "preserve and resend phase on all assistant messages —
-#     dropping it can degrade performance."  Required for prefix cache hits.
-#   * ``finish_reason``: informational; cheap to keep so transcripts replay
-#     identically across CLI and gateway.
-_ASSISTANT_REPLAY_FIELDS: tuple[str, ...] = (
-    "reasoning",
-    "reasoning_content",
-    "reasoning_details",
-    "codex_reasoning_items",
-    "codex_message_items",
-    "finish_reason",
+# --- Resume/replay helpers (extracted to gateway.resume_replay_helpers) ---
+from gateway.resume_replay_helpers import (  # noqa: E402
+    _ASSISTANT_REPLAY_FIELDS,
+    _build_replay_entry,
+    build_resume_recovery_note,
 )
-
-
-def _build_replay_entry(
-    role: str,
-    content: Any,
-    msg: Dict[str, Any],
-    preserve_timestamp: bool = False,
-) -> Dict[str, Any]:
-    """Build a replay entry for a non-tool-calling message, preserving the
-    assistant fields the agent's API builders rely on for multi-turn fidelity.
-
-    Lifted out of the inline ``run_sync`` closure so the field whitelist can
-    be unit-tested in isolation.  Mirrors the ``_ASSISTANT_REPLAY_FIELDS``
-    contract above.
-
-    ``preserve_timestamp``: when True, copy the source row's ``timestamp``
-    onto the replay entry. Currently only user messages need this — the
-    stale-dangerous-confirmation stripper in ``agent/replay_cleanup.py``
-    reads the timestamp to decide whether a confirmation is too old to
-    replay safely.  Assistant/tool messages are not timestamp-stripped in
-    the same way, so we keep the existing default of dropping it.
-
-    Empty values: most fields are dropped when falsy (matching the original
-    PR #2974 behaviour) since an empty list/string for those carries no
-    information.  The exception is ``reasoning_content``: DeepSeek/Kimi
-    thinking-mode replay treats an empty string as a meaningful sentinel
-    that ``_copy_reasoning_content_for_api`` upgrades to a single space.
-    Dropping it here would make the gateway send no ``reasoning_content``
-    at all on the next turn, which can cause HTTP 400 from strict thinking
-    providers.
-    """
-    entry: Dict[str, Any] = {"role": role, "content": content}
-    # api_content sidecar (persist-what-you-send, prompt-cache stability):
-    # forward the exact bytes previously sent to the API for this message so
-    # the agent's api_messages build can substitute them and keep the request
-    # prefix byte-stable across turns. Forward ONLY when this replay pipeline
-    # did not rewrite the content (timestamp injection, auto-continue strip,
-    # mirror prefix): a rewritten clean content means the pipeline decided
-    # different bytes must replay — resending the stored sidecar would
-    # reintroduce exactly what was stripped. Dropping it costs one cache
-    # boundary; resending stripped noise is a behavior regression.
-    _sidecar = msg.get("api_content")
-    if (
-        role in ("user", "assistant")
-        and isinstance(_sidecar, str)
-        and _sidecar
-        and content == msg.get("content")
-    ):
-        entry["api_content"] = _sidecar
-    if role == "assistant":
-        for _rkey in _ASSISTANT_REPLAY_FIELDS:
-            if _rkey not in msg:
-                continue
-            _rval = msg.get(_rkey)
-            if _rkey == "reasoning_content":
-                # Preserve empty-string sentinel for thinking-mode replay.
-                if _rval is None:
-                    continue
-            elif not _rval:
-                continue
-            entry[_rkey] = _rval
-    if preserve_timestamp:
-        ts = msg.get("timestamp")
-        if ts:
-            entry["timestamp"] = ts
-    return entry
 
 
 _TELEGRAM_OBSERVED_CONTEXT_PROMPT_MARKER = "observed Telegram group context"
